refactoring(15): extract settingsConstants from SettingsPage
Fourth slice of req-15 (multi-PR). SettingsPage.tsx 1123 -> 1053 lines. It is a
heavy 38-useState form container, so NO state-bearing UI was extracted; instead
the self-contained constants/types/helper block (provider/model + Letta embedding
tables, CUSTOM_HANDLE_VALUE, LettaModelEntry/LettaProviderTable types, the pure
_splitHandle helper) moved VERBATIM into components/Settings/settingsConstants.ts
(pure data/logic, no JSX). Container imports them; byte-identical (modulo export
kw); dropped the unused LettaProviderTable import.

Checks: typecheck 0 non-test errors; build ok; vitest 4 passed / 2 failed
(pre-existing brittle PaperBench tests). req-15 file RETAINED. Remaining:
DetailPanel (monolith — needs higher-risk prop-contract refactor, not verbatim).
