+++ ari-core/ari/viz/frontend/src/components/README.md
@@ -60,6 +60,7 @@ Feature-grouped React components — one subdirectory per dashboard page plus sh
 - `Settings/` — dashboard/run configuration page.
   - `README.md` — Settings index.
   - `index.ts` — barrel re-export.
+  - `settingsConstants.ts` — provider/Letta model tables + _splitHandle helper (extracted from SettingsPage in req 15).
   - `SettingsPage.tsx` — settings view.
 - `Tree/` — BFTS tree page (search tree, detail panel, file browser).
   - `README.md` — Tree index.
